server/requirements.txt
appdirs==1.4.4
asgiref==3.3.1
CacheControl==0.12.6
certifi==2020.11.8
chardet==3.0.4
colorama==0.4.4
colorclass==2.2.0
contextlib2==0.6.0.post1
distlib==0.3.1
distro==1.5.0
Django==3.1.3
django-cors-headers==3.5.0
django-filter==2.4.0
django-mysql==3.9.0
djangorestframework==3.12.2
docopt==0.6.2
html5lib==1.1
idna==2.10
ipaddr==2.2.0
lockfile==0.12.2
Markdown==3.3.3
msgpack==1.0.0
mysqlclient==2.0.1
packaging==20.4
pep517==0.9.1
pip-upgrader==1.4.15
progress==1.5
pyparsing==2.4.7
pytoml==0.1.21
pytz==2020.4
requests==2.25.0
retrying==1.3.3
six==1.15.0
sqlparse==0.4.1
terminaltables==3.1.0
toml==0.10.2
urllib3==1.26.2
webencodings==0.5.1
